+++ src/components/login/utils/components/LoginKeyframes.tsx
@@ -1,0 +1,26 @@
+export function LoginKeyframes() {
+  return (
+    <style>{`
+      @keyframes fadeUp {
+        from { opacity: 0; transform: translateY(20px); }
+        to { opacity: 1; transform: translateY(0); }
+      }
+      @keyframes shapeFloatTL {
+        0%, 100% { transform: translate(0, 0); }
+        50% { transform: translate(10px, 15px); }
+      }
+      @keyframes shapeFloatTR {
+        0%, 100% { transform: rotate(0deg); }
+        50% { transform: rotate(3deg); }
+      }
+      @keyframes rotate {
+        from { transform: rotate(0deg); }
+        to { transform: rotate(360deg); }
+      }
+      @keyframes pathDraw {
+        from { stroke-dasharray: 1000; stroke-dashoffset: 1000; }
+        to { stroke-dasharray: 1000; stroke-dashoffset: 0; }
+      }
+    `}</style>
+  );
+}
