+++ src/components/login/utils/getErrorMessage.ts
@@ -1,0 +1,10 @@
+export function getErrorMessage(err: unknown) {
+  if (err instanceof Error) return err.message;
+  if (typeof err === "string") return err;
+
+  if (typeof err === "object" && err !== null && "message" in err) {
+    const msg = (err as { message?: unknown }).message;
+    if (typeof msg === "string") return msg;
+  }
+  return "로그인에 실패했어요. 다시 시도해주세요.";
+}
